Adding getNbFrames to image_providers + adding access to ImageProviders in monitoring manager

+++ src/hl_monitoring/replay_image_provider.cpp
@@ -24,7 +24,7 @@ ReplayImageProvider::ReplayImageProvider(const std::string & video_path,
   loadMetaInformation(meta_information_path);
 }
 
-int ReplayImageProvider::getNbFrames() const {
+size_t ReplayImageProvider::getNbFrames() const {
   return nb_frames;
 }
